updating modal images
